refactor(DPGO-19): ad passw handlers
